Create auth-protected pages and endpoints
Pages are protected by custom auth middleware and endpoints are protected by an auth "guard". At start up, we generate a random authorization token and print it to the console. To log in, users must provide the token to the `login` endpoint. In the UI, this can be accomplished with a post `fetch` request setting `redirect` to "follow" and manually handling the redirect:

```javascript
if (resp.redirected) {
  window.location.href = resp.url;
}
```

We use built-in cookie-based session middleware to keep track of whether users are logged in or not. A user is authenticated if their request contains a session with `token=AUTH_TOKEN`.

In `DEV_MODE`, use secret token "test" to authorize.