Create asgi.py
This commit relocates `app` from `__init__.py` to `asgi.py`. For a minimal application, `__init__.py` is a reasonable location (and one of the locations where Litestar automatically checks for an `app`), but for a more complex application where we are importing objects from `src/app`, this location becomes problematic.